chore: improve logic
